Time: 2 ms (99.11%) | Memory: 10 MB (99.07%) - LeetSync

+++ 518-coin-change-ii/coin-change-ii.cpp
@@ -2,9 +2,7 @@ class Solution {
 public:
     int change(int amount, vector<int>& coins) {
         int n=coins.size();
-        // vector<vector<int>>dp(n+1,vector<int>(amount+1,0));
-        // for(int i=0;i<=n;++i) dp[i][0]=1;
-        vector<long  unsigned>dp(amount+1,0);
+        vector< unsigned>dp(amount+1,0);
         dp[0]=1;
         for(int i=0;i<n;++i){
             for(int am=coins[i];am<=amount;++am){
